requirements.txt
asgiref==3.9.1
Django==5.0
django-simple-history==3.10.1
sqlparse==0.5.3
